Move DeleteVideoDialog and WatchedTimesDialogs into VideoDialogs folder
